working example with test image
This example does let through a lot of pixel possibilities, this is because the color pattern. The only time it will return false when the X is all the way to the right (because there aren't any pixels to the right of it which it needs to pass the variant). Same story with the Y when it is at the bottom